[fircg] Lower to saturated fp to int conversions
fptosi and fptoui llvm instructions become poison during optimization
when the conversion cannot be performed, such as a narrowing conversion.

The standard mandates that a real converted to an int must be converted
to the largest integer that does not exceed the magnitude of the original
and keeps the same sign. The saturated floating point conversions match
these semantics more closely than the regular conversion instructions.

A potential key difference is the handling of NaNs- NaNs become 0 instead
of INT_MIN or INT_MAX, as with other compilers.

This change does not include a test for fptoui because I do not know if
that is possible from fortran source.