[flang] Add represention of volatile references
Flang currently lacks support for volatile variables. For some cases, the
compiler produces TODO error messages and others are ignored. Some of our tests
are like the example from C.4 Clause 8 notes: The VOLATILE attribute (8.5.20)
and require volatile variables.

This change is a minimal draft of support for volatility in Fortran. This
misses several important features that will be required for full support,
such as volatile boxes and other reference-like types. This commit only
supports volatility for !fir.ref<T> types and is the minimum needed to
get end-to-end examples working.