[mlir][python] Wrappers for scf.index_switch
The C++ index switch op has utilies for getCaseBlock(int i)
and getDefaultBlock(), so these have been added.

Optional builder args have been added for the default case
and each switch case.

The list comprehensions for accessing case regions are due
to what appears to be a bug in RegionSequence; using a comprehension
with explicit indices circumvents this.
The same paradigm is used for get_case_block(i: int), but this
is unavoidable.